hack: inject content-type into clients
