+++ pkg/cliplugins/workspace/plugin/get.go
@@ -13,6 +13,7 @@ import (
 	tenancyv1alpha1 "github.com/faroshq/faros/pkg/apis/tenancy/v1alpha1"
 	"github.com/faroshq/faros/pkg/cliplugins/base"
 	utilprint "github.com/faroshq/faros/pkg/util/print"
+	"github.com/faroshq/faros/pkg/util/rest"
 )
 
 // GetOptions contains options for configuring faros
@@ -79,7 +80,7 @@ func (o *GetOptions) Run(ctx context.Context) error {
 
 	// Check organization exists
 	organizations := tenancyv1alpha1.OrganizationList{}
-	err = farosClient.RESTClient().Get().AbsPath(o.TenantOrganizationsAPI).Do(ctx).Into(&organizations)
+	err = rest.ContentTypeJSON(farosClient.RESTClient().Get()).AbsPath(o.TenantOrganizationsAPI).Do(ctx).Into(&organizations)
 	if err != nil {
 		return err
 	}
@@ -107,7 +108,7 @@ func (o *GetOptions) Run(ctx context.Context) error {
 		return err
 	}
 	workspaces := &tenancyv1alpha1.WorkspaceList{}
-	err = farosClient.RESTClient().Get().AbsPath(path).Do(ctx).Into(workspaces)
+	err = rest.ContentTypeJSON(farosClient.RESTClient().Get()).AbsPath(path).Do(ctx).Into(workspaces)
 	if err != nil {
 		return err
 	}
