+++ pkg/cliplugins/workspace/plugin/use.go
@@ -14,6 +14,7 @@ import (
 
 	tenancyv1alpha1 "github.com/faroshq/faros/pkg/apis/tenancy/v1alpha1"
 	"github.com/faroshq/faros/pkg/cliplugins/base"
+	"github.com/faroshq/faros/pkg/util/rest"
 )
 
 // UseOptions contains options for configuring faros
@@ -74,7 +75,7 @@ func (o *UseOptions) Run(ctx context.Context) error {
 
 	// Check organization exists
 	organizations := tenancyv1alpha1.OrganizationList{}
-	err = farosClient.RESTClient().Get().AbsPath(o.TenantOrganizationsAPI).Do(ctx).Into(&organizations)
+	err = rest.ContentTypeJSON(farosClient.RESTClient().Get()).AbsPath(o.TenantOrganizationsAPI).Do(ctx).Into(&organizations)
 	if err != nil {
 		return err
 	}
@@ -110,7 +111,7 @@ func (o *UseOptions) Run(ctx context.Context) error {
 
 	workspace := &tenancyv1alpha1.Workspace{}
 	if o.Name != tenancyv1alpha1.KubeConfigAuthKey {
-		err = farosClient.RESTClient().Get().AbsPath(path).Do(ctx).Into(workspace)
+		err = rest.ContentTypeJSON(farosClient.RESTClient().Get()).AbsPath(path).Do(ctx).Into(workspace)
 		if err != nil {
 			return err
 		}
