+++ pkg/util/roundtripper/rountripper.go
@@ -1,0 +1,11 @@
+package roundtripper
+
+import (
+	"net/http"
+)
+
+type RoundTripperFunc func(*http.Request) (*http.Response, error)
+
+func (f RoundTripperFunc) RoundTrip(req *http.Request) (*http.Response, error) {
+	return f(req)
+}
